major sizeof(int)=4 work

+++ binutils/asm/pass2.c
@@ -216,13 +216,13 @@ short pass2( void )
 
 void write_string_table( void )
 {
-    long t = 0;
+    int t = 0;  /* COFF uses 32-bit size */
 
     if( string_table_index > 4 )
-      t = string_table_index;
+      t = (int)string_table_index;
 
     fseek( obj_fd, 0L, 2 );
-    FWRITE( &t, sizeof(long), 1, obj_fd );
+    FWRITE( &t, sizeof(int), 1, obj_fd );  /* Write 4 bytes, not 8 */
 #ifdef MSDOS
     FWRITE( &string_table[4], sizeof(char), (size_t) (string_table_index - 4), obj_fd );
 #else
